refactor: MyProjectPosts 관련 api, type 분리
